reset styling and architecture; mobile 1st
