+++ src/components/Header.scss
@@ -1,0 +1,1 @@
+.header {}
