+++ src/components/IndexIntro.scss
@@ -1,0 +1,43 @@
+.home-container {
+  border: 1px solid yellow;
+}
+
+.home-container {
+  // border: 1px solid yellow;
+  // width: 100%;
+  // display:flex;
+  // justify-content: center;
+  // align-items: center;
+  // position:relative;
+  // height:140vh;
+
+  // @media screen and (max-width:320px){
+  //   width:100%;
+  //   display:block;
+  // }
+  // @media screen and (max-width:375px){
+  //   width:100%;
+  //   display:block;
+  // }
+  // @media screen and (max-width:425px){
+  //   width:100%;
+  //   height:145vh;
+  //   display:block;
+  // }
+  // @media screen and (max-width:768px){
+  //   width:100%;
+  //   height:200vh;
+  //   display:block;
+  // }
+  // @media screen and (max-width:1024px){
+  //   width:100%;
+  //   height:200vh;
+  //   display:block;
+  // }
+  // @media screen and (max-width:1440px){      
+  //   height:100vh;
+  // }
+  // @media screen and (max-width:2560px){      
+  //   height:55vh;
+  // }
+}
