--- src/components/layout.js
@@ -1,22 +1,0 @@
-import * as React from "react"
-import { GlobalStyle } from "./styles/GlobalStyles"
-import Header from "./Header"
-
-const Layout = ({ children }) => {
- 
-  
-  return (
-    <>
-      <GlobalStyle />
-      <Header  />
-     
-        <main>{children}</main>
-        
-          
-    </>
-  )
-}
-
-
-
-export default Layout
