--- src/components/styles/GlobalStyles.js
@@ -1,13 +1,0 @@
-import { createGlobalStyle } from "styled-components";
-
-export const GlobalStyle = createGlobalStyle`
-*{
-    box-sizing:border-box ;
-    background-color:black;
-    margin: 0;
-    padding: 0;
-    max-width: 98.8vw !important;
-}
-
-`
-
