+++ src/pages/styles.global.scss
@@ -1,0 +1,25 @@
+@import '../styles.config.scss';
+
+* {
+  box-sizing: border-box;
+  margin: 0;
+  max-width: 100vw;
+  padding: 0;
+}
+
+body {
+  background-color: black;
+  font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', 'Roboto', 'Oxygen', 'Ubuntu', 'Cantarell', 'Fira Sans', 'Droid Sans', 'Helvetica Neue', sans-serif;
+  -webkit-font-smoothing: antialiased;
+  -moz-osx-font-smoothing: grayscale;
+}
+
+// mobile first
+
+.psyoptions {
+  margin: 8px;
+
+  @media screen and (min-width: 768px) {
+    margin: 0;
+  }
+}
