+++ src/styles.config.scss
@@ -1,0 +1,1 @@
+// $test: 0;
